\name{cwhplot-package}
\alias{cwhplot-package}
\docType{package}
\title{
cwh's plot functions
}
\description{
Miscellaneous plot functions for general use
}
\details{
\tabular{ll}{
Package: \tab cwhplot\cr
Version: \tab 4.2.0\cr
Date: \tab 2006-01-26\cr
Depends: \tab R (>= 2.0.0)\cr
BundleDescription: \tab Miscellaneous functions for general use\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.wsl.ch/staff/christian.hoffmann/\cr
}

Index:
\preformatted{
T3plot                  T3plot
lowess.bygroup          Plot data in groups, each with lowess or loess
                        smoothing, may be used in v
lpr                     Print an object or plot
mult.fig.p              Plot Setup for MULTiple FIGures, incl. Main
                        Title
p.screeplot.princomp    Plot screeplot
panel.cor               Alternative panel functions for lattice plots
plotSymbols             Plot symbols, colours, and allow to choose
plt                     Plot depending on switch, Create multiple plots
                        with title and time stamp
pltSplomT               Plot histograms, scatterplot matrix
}
}
\author{
Christian W. Hoffmann <christian.hoffmann@wsl.ch>

Maintainer: Christian W. Hoffmann <christian.hoffmann@wsl.ch>
}
\keyword{ package }
